requirements.txt
requests==2.32.3
openai==1.53.0
python-dotenv==1.0.1
httpx==0.27.2

